@@ -1,63 +1,59 @@
 buildscript {
     repositories {
-//        mavenCentral()
         jcenter()
+        google()
     }
 
     dependencies {
-        classpath 'com.android.tools.build:gradle:2.3.0'
+        classpath 'com.android.tools.build:gradle:3.5.0'
     }
 }
 
 apply plugin: 'com.android.application'
 
 android {
-    compileSdkVersion 23
-    buildToolsVersion '25.0.0'
-
+    compileSdkVersion 29
     defaultConfig {
         applicationId "com.hekkelman.keylocker"
         minSdkVersion 15
-        targetSdkVersion 23
+        targetSdkVersion 29
         versionCode 1
         versionName "1.0"
     }
-
     repositories {
-//        maven { url 'http://repo1.maven.org/maven2' }
         jcenter()
     }
-
     buildTypes {
         release {
             minifyEnabled false
             proguardFiles getDefaultProguardFile('proguard-android.txt'), 'proguard-rules.pro'
         }
     }
+    buildToolsVersion '28.0.3'
+    productFlavors {
+    }
 }
 
 dependencies {
-    compile fileTree(dir: 'libs', include: ['*.jar'])
-    testCompile 'junit:junit:4.12'
-
-    compile('org.simpleframework:simple-xml:2.7.1') {
+    implementation fileTree(include: ['*.jar'], dir: 'libs')
+    //    testCompile 'junit:junit:4.12'
+    annotationProcessor 'com.jakewharton:butterknife:7.0.1'
+    implementation 'com.android.support:support-v4'
+    implementation('org.simpleframework:simple-xml:2.7.1') {
         exclude module: 'stax'
         exclude module: 'stax-api'
         exclude module: 'xpp3'
     }
-
-    compile 'com.android.support:appcompat-v7:23.1.1'
-    compile 'com.android.support:cardview-v7:23.1.1'
-    compile 'com.android.support:design:23.1.1'
-
+    implementation 'androidx.appcompat:appcompat:1.1.0'
+    implementation 'com.android.support:cardview-v7'
+    implementation 'com.android.support:design'
     // Include the sdk as a dependency
-    compile ('com.onedrive.sdk:onedrive-sdk-android:1.2+') {
+    implementation('com.onedrive.sdk:onedrive-sdk-android:1.3+') {
         transitive = false
     }
-
-    compile 'com.google.code.gson:gson:2.3.1'
-    compile 'com.microsoft.services.msa:msa-auth:0.8.+'
-    compile 'com.microsoft.aad:adal:1.1.+'
-    compile 'com.android.support:support-v4:23.1.1'
-    compile 'com.jakewharton:butterknife:7.0.1'
+    implementation 'com.google.code.gson:gson:2.8.2'
+    implementation 'com.microsoft.services.msa:msa-auth:0.8.+'
+    implementation 'com.microsoft.aad:adal:1.1.+'
+    implementation 'com.android.support:support-v4'
+    implementation 'com.jakewharton:butterknife:7.0.1'
 }
