@@ -8,9 +8,11 @@ import android.content.DialogInterface;
 import android.content.pm.PackageManager;
 import android.os.AsyncTask;
 import android.os.Environment;
-import android.support.v4.content.ContextCompat;
+
+//import android.support.v4.content.ContextCompat;
 import android.util.Log;
 
+import androidx.core.content.ContextCompat;
 import com.hekkelman.keylocker.datamodel.InvalidPasswordException;
 import com.hekkelman.keylocker.datamodel.KeyDb;
 import com.onedrive.sdk.extensions.IOneDriveClient;
@@ -20,7 +22,7 @@ import java.io.ByteArrayOutputStream;
 import java.io.File;
 import java.io.InputStream;
 
-import static android.support.v7.widget.StaggeredGridLayoutManager.TAG;
+//import static android.support.v7.widget.StaggeredGridLayoutManager.TAG;
 
 /**
  * Created by maarten on 27-11-15.
